+++ DomainModel/Patient.cs
@@ -8,7 +8,7 @@ namespace DomainModel
 {
     public class Patient : Person
     {
-        public int SocialSecurityNumber { get; set; }
+        public long SocialSecurityNumber { get; set; }
         public virtual ICollection<Appointment> Appointments { get; set; }
     }
 }
